# docs/openapi.yaml
openapi: 3.0.3
info:
  title: sing-box-agent API
  version: 1.2.0
  description: |
    A lightweight Go service that embeds sing-box as a library and provides a REST API for remote management.
    Designed to replace 3x-ui panels and SSH-based config management while maintaining multi-server architecture.

    **Key Features:**
    - Runtime user management without process restarts
    - Full protocol support via sing-box (VLESS, Hysteria2, ShadowTLS, TUIC, etc.)
    - Multi-server architecture — one agent per VPN server, controlled centrally
    - Prometheus metrics export for Grafana dashboards
    - Zero-dependency deployment (single binary)
    - State consistency — API is source of truth, agent applies desired state

  contact:
    name: sing-box-agent
  license:
    name: MIT
servers:
  - url: http://localhost:8080
    description: Local development
  - url: https://agent.example.com
    description: Production server
security:
  - BearerAuth: []
  - RequestSigning: []
...
paths:
...
  /subscription/generate:
    post:
      tags:
        - Subscription
      summary: Generate client config
      description: Generate client configuration for a user
      operationId: generateSubscription
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - subId
                - inboundTag
              properties:
                subId:
                  type: string
                  description: Subscription ID
                inboundTag:
                  type: string
                  description: Inbound tag
                format:
                  type: string
                  enum: [vless, vmess, trojan, hysteria2, tuic]
                  default: vless
                  description: Output format
      responses:
        '200':
          description: Configuration generated successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/SubscriptionConfig'
  /subscription/{subId}:
    get:
      tags:
        - Subscription
      summary: Get subscription
      description: Retrieve subscription for a user
      operationId: getSubscription
      parameters:
        - name: subId
          in: path
          required: true
          schema:
            type: string
          description: Subscription ID
      responses:
        '200':
          description: Subscription details
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Subscription'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Bearer token for authentication (minimum 32 characters)
    RequestSigning:
      type: apiKey
      in: header
      name: X-Signature
      description: |
        HMAC-SHA256 signature for request authentication

        **Required Headers:**
        - X-Signature: HMAC-SHA256(canonical_string, secret)
        - X-Timestamp: Unix seconds (reject if |server_time - timestamp| > 300s)
        - X-Nonce: UUID v4 (stored in LRU cache, 24h TTL, 10k max entries)

        **Canonicalization:**
        ```
        canonical_string = "{nonce}\\n{timestamp}\\n{method}\\n{path}\\n{body_hash}"
        body_hash = SHA256(request_body).hexdigest()  // empty string if no body
        signature = HMAC-SHA256(canonical_string, secret).hexdigest()
        ```
  schemas:
    SuccessResponse:
      type: object
      required:
        - success
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
          description: Response payload
    ErrorResponse:
      type: object
      required:
        - success
        - error
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          required:
            - code
            - message
          properties:
            code:
              type: string
              enum:
                - INVALID_REQUEST
                - UNAUTHORIZED
                - REPLAY_DETECTED
                - INBOUND_NOT_FOUND
                - USER_NOT_FOUND
                - USER_EXISTS
                - IDEMPOTENCY_CONFLICT
                - VERSION_CONFLICT
                - SYNC_FAILED
                - SERVICE_UNAVAILABLE
              description: Error code
            message:
              type: string
              description: Human-readable error message
...
    SubscriptionConfig:
      type: object
      required:
        - config
        - link
      properties:
        config:
          type: object
          description: Client configuration object
        link:
          type: string
          description: Connection link (e.g., vless://...)
    Subscription:
      type: object
      required:
        - subId
        - configs
      properties:
        subId:
          type: string
          description: Subscription ID
        configs:
          type: array
          items:
            type: object
            description: Available configurations for different protocols
        updated_at:
          type: string
          format: date-time
          description: Last update timestamp
...